refactor(luceneQuery): move the form shape helpers next to the query parser
